% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aucs.bouldr.R
\name{aucs}
\alias{aucs}
\title{Build a table of AUCs for each ROC}
\usage{
aucs(rocbag)
}
\arguments{
\item{rocbag}{A \code{bouldr}-class object}
}
\value{
A data frame with the AUCs
}
\description{
Build a table of AUCs for each ROC
}
