% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouldr.R
\name{bouldr}
\alias{bouldr}
\title{Run ROCs and pairwise comparisons using desired test}
\usage{
bouldr(formula, data, levels, direction, test = "delong", ...)
}
\arguments{
\item{formula}{A specification of the variables of form: response ~ predictor + (grouping_var) + (faceting_var).
Arguments in parentheses are optional.}

\item{data}{A data frame containing the response and predictor vars}

\item{levels}{The levels of the response variable as a character vector (e.g., c("yes","no"))}

\item{direction}{Either "<" or ">", indicating which level is the positive case}

\item{test}{Which test should be used to compare ROCs (optional). Default is delong}

\item{...}{Additional arguments}
}
\value{
An object of class \code{bouldr}' containing ROCs and tests.
}
\description{
This function takes a data frame and a formula and runs nested ROCs
as appropriate. Also required are the levels of the response variable
and the direction of the cases (i.e., which level is the positive case).
Providing an argument for test will change the test used to compare
ROC curves.
}
\examples{
testing <- generate_data(10,Diagnosis = c("Depression","Anxiety","ADHD"),
        Measure = c("A","B"), Informant = c("Self","Parent", "Teacher"))
single <- bouldr(data = testing,
formula = Outcome ~ Score,
test = 'delong',
levels = c('no','yes'),
direction = "<")
aucs(single)
}
