% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilrs.ntiles.bouldr.R
\name{dlrs.ntiles}
\alias{dlrs.ntiles}
\title{A Diagnostic Likelihood Ratio (DiLR) calculator for n-tiles}
\usage{
dlrs.ntiles(n, ...)
}
\arguments{
\item{n}{(numeric) the number of cut points to include}

\item{...}{Additional arguments passed to \code{dilrs()}}
}
\value{
a data frame containing the classification statistics for each cut point
}
\description{
Computes Diagnostic Likelihood Ratios (DiLRS) for a specified quantiles of
the continuous variable. Also outputs additional
classification statistics (true/false positives/negatives).
}
