% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate a test data frame with 6 nested variables}
\usage{
generate_data(
  obs.per,
  Diagnosis = c("Depression", "Anxiety"),
  Measure = c("A", "B", "C", "D", "E"),
  Informant = c("Parent", "Teacher", "Self")
)
}
\arguments{
\item{obs.per}{Number of observations per cell}

\item{Diagnosis}{A character vector of names for this grouping variable}

\item{Measure}{A character vector of names for this grouping variable}

\item{Informant}{A character vector of names for this grouping variable}
}
\value{
A data frame that can be used in bouldr
}
\description{
ID: participant ID
Diagnosis: the type of diagnosis
Measure: the form used to measure symptoms
Informant: the person taking the form
Outcome: whether the diagnosis is present
Score: the score on the measure
}
