% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bouldr.R
\name{summary.bouldr}
\alias{summary.bouldr}
\title{Summarises the \code{bouldr} output (by passing to print)}
\usage{
\method{summary}{bouldr}(object, ...)
}
\arguments{
\item{object}{A \code{bouldr} object}

\item{...}{additional arguments}
}
\description{
Summarises the \code{bouldr} output (by passing to print)
}
