% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testData.R
\docType{data}
\name{testData}
\alias{testData}
\title{Test data for bouldr}
\format{
A data frame with 3000 rows and 6 columns.
}
\usage{
testData
}
\description{
A test data set, generated using generate_data()
}
\keyword{datasets}
