% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.bouldr.R
\name{tests}
\alias{tests}
\title{Convenience function to pull out the table of test results}
\usage{
tests(rocbag)
}
\arguments{
\item{rocbag}{A bouldr-class object}
}
\value{
A data frame containing the test information
}
\description{
Convenience function to pull out the table of test results
}
