% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilrs.bouldr.R
\name{dlrs}
\alias{dlrs}
\title{A Diagnostic Likelihood Ratio (DiLR) calculator}
\usage{
dlrs(x, scores, outcomes, positive, ...)
}
\arguments{
\item{x}{a data frame containing the relevant variables}

\item{scores}{(character) the name of the continuous variable}

\item{outcomes}{(character) the name of the classification variable}

\item{positive}{(string or numeric) the positive case in \code{outcomes}}

\item{...}{Additional arguments including \code{n.cuts} or \code{cut} (one of these is required). Supplying \code{cut} computes dlrs for a specific cut score. Supplying \code{n.cuts} computes it for a range of cut scores based on quantiles of the score variable.}
}
\value{
a numeric vector containing the classification statistics
}
\description{
Computes Diagnostic Likelihood Ratios (DiLRS) for a specified cut-point
of the continuous variable. Also outputs additional
classification statistics (true/false positives/negatives).
}
