% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Title: Generate a test data frame for use with \code{bouldr}}
\usage{
generate_data(
  obs.per,
  Diagnosis = c("Depression", "Anxiety"),
  Measure = c("A", "B", "C", "D", "E"),
  Informant = c("Parent", "Teacher", "Self")
)
}
\arguments{
\item{obs.per}{Number of observations per cell. Total \emph{N} would be this number times the number of options in the other three variables.}

\item{Diagnosis}{A character vector of names for this grouping variable}

\item{Measure}{A character vector of names for this grouping variable}

\item{Informant}{A character vector of names for this grouping variable}
}
\value{
A data frame that can be used in \code{bouldr}
}
\description{
Description: Generates a data frame of the correct format to be used
with \code{bouldr}.
}
\details{
By default, the output data frame frame contains the following variables:
\itemize{
\item ID: participant ID
\item Diagnosis: the type of diagnosis. One or more -- default 2: "Depression" and "Anxiety"
\item Measure: the form used to measure symptoms. One or more -- default 5: (A through E)
\item Informant: the person taking the form. One or more -- default 3: "Parent", "Teacher" and "Self"
\item Outcome: whether the diagnosis is present. Must be dichotomous -- default: "yes" and "no"
\item Score: the score on the measure. A (continuous) numeric variable.
}

Note that these variable names are arbitrary; they have been
chosen here to reflect common use-cases.
}
\examples{
generate_data(10)
}
